refactor code and toss unused functions
