+++ app/audits/page.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 
 import { Loader } from "@/components/Loader";
-import { getAudits } from "@/lib/actions/audits";
+import { getAudits } from "@/actions/audits";
 import Audits from "@/components/screens/audits";
 
 const Fetcher = async (): Promise<JSX.Element> => {
