+++ components/Dropdown/Content/networks.tsx
@@ -3,7 +3,7 @@ import { useSwitchChain, useAccount, Register } from "wagmi";
 
 import { Icon } from "@/components/Icon";
 import * as Card from "@/components/Card";
-import { ChainPresets } from "@/lib/constants/chains";
+import { ChainPresets } from "@/lib/constants";
 import { Column, HoverItem } from "@/components/Box";
 import { Check } from "@/assets";
 import { cn } from "@/lib/utils";
