--- hooks/useIsMounted.tsx
@@ -1,11 +1,0 @@
-import { useState, useEffect } from "react";
-
-export const useIsMounted = (): boolean => {
-  const [mounted, setMounted] = useState(false);
-  useEffect(() => {
-    setMounted(true);
-    return () => setMounted(false);
-  }, []);
-
-  return mounted;
-};
