+++ lib/types.ts
@@ -1,4 +1,4 @@
-import { Users } from "@prisma/client";
+import { Prisma, Users } from "@prisma/client";
 import { Address } from "viem";
 
 export type LeaderboardI = {
@@ -74,3 +74,70 @@ export type UserStateI = {
   isFetched: boolean;
   isFetchedAfterMount: boolean;
 };
+
+export interface UserWithCount extends Users {
+  stats: {
+    valuePotential: number;
+    valueComplete: number;
+    numActive: number;
+    numComplete: number;
+  };
+}
+
+export type AuditViewDetailedI = Prisma.AuditsGetPayload<{
+  include: {
+    auditee: true;
+    auditors: {
+      include: {
+        user: true;
+      };
+    };
+  };
+}>;
+
+export type AuditViewI = Prisma.AuditsGetPayload<{
+  include: {
+    auditee: true;
+    auditors: {
+      select: {
+        user: true;
+      };
+    };
+  };
+}>;
+
+export type UserAuditsI = Prisma.UsersGetPayload<{
+  select: {
+    auditees: {
+      select: {
+        auditors: {
+          include: {
+            user: true;
+          };
+        };
+      };
+    };
+  };
+}>;
+
+export interface UserStats {
+  moneyPaid: number;
+  moneyEarned: number;
+  numAuditsCreated: number;
+  numAuditsAudited: number;
+}
+
+interface GenericSuccess<T> {
+  success: boolean;
+  data: T;
+  error?: never;
+  validationErrors?: never;
+}
+interface GenericFailure {
+  success: boolean;
+  data?: never;
+  error: string;
+  validationErrors?: Record<string, string>;
+}
+
+export type GenericUpdateI<T> = GenericSuccess<T> | GenericFailure;
