+++ providers/modal/context.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+import { ModalStateI } from "@/lib/types";
+
+const ModalContext = createContext<ModalStateI>({
+  toggleOpen: () => {},
+  setContent: () => {},
+});
+
+export default ModalContext;
